% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDumbbell.R
\name{plotDumbbell}
\alias{plotDumbbell}
\title{Dumbbell plot for comparing two logFCs}
\usage{
plotDumbbell(gene.list, table.1, table.2, label.1, label.2, plot.title,
  plot.subtitle, do.return = F)
}
\arguments{
\item{gene.list}{List of genes to plot}

\item{table.1}{EdgeR table (topTags output) for the first comparison}

\item{table.2}{EdgeR table (topTags output) for the second comparison}

\item{label.1}{Label for the first comparison}

\item{label.2}{Label for the second comparison}

\item{plot.title}{Title for the plot}

\item{plot.subtitle}{Subtitle for the plot}

\item{do.return}{Whether to return a ggplot object}
}
\description{
plotDumbbell takes as input a list of genes and two tables produced by edgeR
to visualize the difference in fold change between the two comparisons. This 
is particularly useful for displaying fold changes in a comparison from an 
interaction term where differences across two factors are being compared.
}
\examples{
WT_table = edgeR::topTags(wt.lrt) \%>\% .$table
KO_table = edgeR::topTags(ko.lrt) \%>\% .$table
plotDumbBell(gene.list = c("Actb", "Gapdh", "Acta2"), table.1 = WT_table, table.2 = KO_table,
             label.1 = "WT", label.2="KO", plot.title = "Comparison of stimulation/nostim logFC",
             plot.subtitle="between WT and KO mice")
}
